initial codebase improvement: move images to wilfred/api
