+++ tests/test_server.py
@@ -8,7 +8,7 @@
 #                                                                  #
 ####################################################################
 
-from wilfred.images import Images
+from wilfred.api.images import Images
 from wilfred.servers import Servers
 from wilfred.docker_conn import docker_client
 from wilfred.server_config import ServerConfig
